+++ src/main/java/insane96mcp/insanesurvivaloverhaul/setup/ISORegistries.java
@@ -16,7 +16,6 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.entity.ai.attributes.Attribute;
-import net.minecraft.world.entity.ai.village.poi.PoiType;
 import net.minecraft.world.item.ArmorMaterial;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -42,7 +41,6 @@ public class ISORegistries {
     public static final DeferredRegister<Attribute> ATTRIBUTES = createRegistry(BuiltInRegistries.ATTRIBUTE);
     public static final DeferredRegister<SoundEvent> SOUND_EVENTS = createRegistry(BuiltInRegistries.SOUND_EVENT);
     public static final DeferredRegister<MobEffect> MOB_EFFECTS = createRegistry(BuiltInRegistries.MOB_EFFECT);
-    public static final DeferredRegister<PoiType> POI_TYPES = createRegistry(BuiltInRegistries.POINT_OF_INTEREST_TYPE);
     @SuppressWarnings("rawtypes")
     private static final DeferredRegister CRITERION_TRIGGERS = createRegistry(BuiltInRegistries.TRIGGER_TYPES);
     public static final DeferredRegister<LootItemConditionType> LOOT_CONDITIONS = createRegistry(BuiltInRegistries.LOOT_CONDITION_TYPE);