+++ changelog.md
@@ -1,3 +1,6 @@
+## Upcoming
+* Fixed missing some lang entries
+
 ## From 1.20.1
 * Combat
   * Unfair one-shot